Support posting errors
Wayland requires that badly behaved clients be disconnected with a fatal
protocol error.  Implement this.  All other compositors behave the same
way.

This also causes the connection to be terminated on any unexpected OCaml
exception.  This indicates either a bug in the callback (a programming
error) or a catastrophic lack of resources.  Servers send either a
wl_display.no_memory or wl_display.implementation error before
disconnecting.  The former is raised for Out_of_memory and
Stack_overflow, while the latter is raised for all other uncaught
exceptions.  The former exceptions are not caught in the callback, as
they are assumed to leave the system in an undefined state.